tried to introduce PersistInAction Flag, but this is not a good way
One thing I hope to achieve is that at the moment, if an action fails, the tree is run again from the beginning, but if it fails, I want the fisherman to do that action again.

The above is the reason why I tried to introduce PersistInAction Flag, but this Flag leaves an issue: after  a caught fish escaped from the fisherman, the fish can no longer be stored.

If I want to accomplish this function, as I mentioned before, it seems to me that introducing another two state "PersistFailure" and "ResetFailure" for node works better compared to just having "Failure" state.